Add updated de-duping logic for next/dynamic
