+++ packages/next-server/lib/utils.ts
@@ -2,7 +2,7 @@ import { format, UrlObject, URLFormatOptions } from 'url'
 import { ServerResponse, IncomingMessage } from 'http'
 import { ComponentType } from 'react'
 import { ParsedUrlQuery } from 'querystring'
-import { ManifestItem } from '../server/get-dynamic-import-bundles'
+import { ManifestItem } from '../server/render'
 import { BaseRouter } from './router/router'
 
 /**
